chore: Update railway.toml, server.js, and package.json
This commit updates the railway.toml, server.js, and package.json files.

In railway.toml, the build builder is set to "NIXPACKS" and the buildCommand is set to "yarn build". In deploy, the startCommand is set to "yarn start", healthcheckPath is set to "/healthz", and healthcheckTimeout is set to 300.

In server.js, the url is now retrieved from the environment variable PUBLIC_URL.

In package.json, the dependencies are reorganized and the pg package is added.